Working on next step.

+++ cuda/muon/decoding/include/MuonRawToHits.cuh
@@ -20,15 +20,15 @@ namespace Muon {
     int cluster_size;
     int region;
 
-    __device__ Hit(HitsSoA* hitsSoA, size_t index);
+    __device__ Hit(HitsSoA* hitsSoA, uint index);
 
     __device__ Hit() {}
   };
 
-  __device__ void setAtIndex(HitsSoA* hitsSoA, size_t index, int tile, float x, float dx, float y, float dy,
+  __device__ void setAtIndex(HitsSoA* hitsSoA, uint index, int tile, float x, float dx, float y, float dy,
       float z, float dz, int uncrossed, unsigned int time, int delta_time, int cluster_size, int region);
 
-  __device__ void setAtIndex(HitsSoA* hitsSoA, size_t index, Hit* hit);
+  __device__ void setAtIndex(HitsSoA* hitsSoA, uint index, Hit* hit);
 
   struct Digit {
     MuonTileID tile;
@@ -44,8 +44,8 @@ namespace Muon {
     MuonTables* muonTables;
     MuonGeometry* muonGeometry;
 
-    __device__ void addCoordsCrossingMap(unsigned int* tileIds, unsigned int* tdcValues, bool* used, size_t startIndex,
-        size_t endIndex, HitsSoA* hitsSoA, int& currentHitIndex) const;
+    __device__ void addCoordsCrossingMap(unsigned int* tileIds, unsigned int* tdcValues, bool* used, uint startIndex,
+        uint endIndex, HitsSoA* hitsSoA, int& currentHitIndex) const;
 
   private:
     __device__ void makeStripLayouts(unsigned int station, unsigned int region, MuonLayout* layouts) const;
